Server/waste_ml-service/requirements.txt
flask
flask-cors
gunicorn
ultralytics
opencv-python-headless
pillow
